+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 0.9.0 stable (2021-03-21)
 
 * Change development status from `beta` to `stable`.
-
+* The parameter `release_date` now also accepts a string in the format `YYYY-MM-DD` besides a datetime object.
 
 ## Version 0.8.0 beta (2020-01-18)